requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
matplotlib==3.8.3
plotly==5.19.0
pydantic==2.6.3
python-dotenv==1.0.1
websockets==12.0
langchain==0.1.11
transformers==4.38.2
torch==2.2.1
